refactor(components.validator): move to webUniverseComponents
ref: OM-174
