+++ .eslintrc.json
@@ -15,7 +15,6 @@
     "angular": true,
     "Chart": true,
     "filesize": true,
-    "ipaddr": true,
     "JustGage": true,
     "moment": true,
     "punycode": true,
