+++ client/app/index.js
@@ -31,8 +31,7 @@ import 'script-loader!@ovh-ux/ovh-utils-angular/bin/ovh-utils-angular.min.js';
 import 'script-loader!@ovh-ux/ovh-utils-angular/lib/core.js';
 import 'script-loader!punycode/punycode.js';
 import 'script-loader!uri.js/src/uri.js';
-import 'script-loader!@bower_components/URIjs/src/URI.min.js';
-import 'script-loader!ipaddr.js/ipaddr.min.js';
+import 'script-loader!URIjs/src/URI.min.js';
 import 'script-loader!filesize/lib/filesize.js';
 import 'ovh-angular-pagination-front';
 import 'script-loader!@bower_components/angular-scroll-glue/src/scrollglue.js';
